+++ niimpy/__init__.py
@@ -2,8 +2,8 @@ from ._version import __version__
 
 from niimpy.reading.database import open, Data1, ALL
 from niimpy.preprocessing.filter import filter_dataframe
-from niimpy.reading.read import read_sqlite, read_sqlite_tables
-from niimpy.reading.read import read_csv, read_csv_string
+from niimpy.reading.sqlite import read_sqlite, read_sqlite_tables
+from niimpy.reading.csv import read_csv, read_csv_string
 from niimpy.preprocessing import sampledata
 from niimpy.preprocessing import util
