+++ niimpy/preprocessing/util.py
@@ -89,7 +89,42 @@ def uninstall_extensions():
     unlink_if_exists(SQLITE3_EXTENSIONS_FILENAME)
 
 
-#TODO: reanme to data.py
+def read_preprocess(df, add_group=None):
+    """Standard preprocessing arguments when reading.
+
+    This is a preprocessing filter which handles some standard arguments
+    when reading files.  This should be considered a private, unstable
+    function.
+
+
+    Parameters
+    ----------
+
+    df: pandas.DataFrame
+
+        Input data frame
+
+    add_group: string, optional
+
+        If given, add a new 'group' column with all values set to this
+        given identifier.
+
+
+    Returns
+    -------
+
+    df: dataframe
+
+        Resulting dataframe (modified in-place if possible, but may also
+        be a copy)
+
+    """
+    if add_group is not None:
+        df['group'] = add_group
+        #df['group'] = df['group'].astype('category')
+        #pd.Categorical(add_group)
+    return df
+
 
 def df_normalize(df, tz=None, old_tz=None):
     """Normalize a df (from sql) before presenting it to the user.
