+++ niimpy/reading/__init__.py
@@ -1,3 +1,4 @@
 from . import mhealth
-from . import read
+from . import csv
 from . import google_takeout
+from . import sqlite
