+++ tests/preprocessing/test_sampledata.py
@@ -6,7 +6,7 @@ import pytest
 from niimpy import config
 
 import niimpy
-from niimpy.reading import read
+from niimpy.reading import csv
 from niimpy.preprocessing import sampledata
 
 TZ = 'Europe/Helsinki'
