+++ tests/reading/test_read.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 import niimpy
-from niimpy.reading import read
+from niimpy.preprocessing import util
 from niimpy import config
 
 TZ = 'Europe/Helsinki'
@@ -9,7 +9,7 @@ TZ = 'Europe/Helsinki'
 def test_read_preprocess_add_group():
     """Test of add_group= option"""
     data = pd.DataFrame({'user': ['u1', 'u2', 'u3'], 'a': [1,2,3], 'b': [4,5,6]})
-    data2 = read._read_preprocess(data, add_group='group1')
+    data2 = util.read_preprocess(data, add_group='group1')
     assert 'group' in data2
     assert (data2['group'] == 'group1').all()
